% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_function.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Regression coefficients}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{A object of class "linreg"}

\item{...}{Arguments belonging to the generic function can be used here}
}
\description{
Prints the regression coefficients as a named vector
}
