% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_function.R
\name{linreg}
\alias{linreg}
\title{Creating the dataframes and calculates the statistics}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A object of class "formula"}

\item{data}{Dataset to use for the linear regression model}
}
\description{
Creating the dataframes and calculates the statistics
}
