% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_function.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Print plots}
\usage{
\method{plot}{linreg}(object)
}
\arguments{
\item{object}{A object of class "linreg"}
}
\description{
Predicted values vs residuals and the Scale-location plots can be created by this function
}
