% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_function.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Print values}
\usage{
\method{pred}{linreg}(object)
}
\arguments{
\item{object}{A object of class "linreg"}
}
\description{
Output the predicted values of the dependent variable
}
