% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_function.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Prints out the residuals of the model}
\usage{
\method{resid}{linreg}(object)
}
\arguments{
\item{object}{A object of class "linreg"}
}
\description{
Prints the residuals
}
