% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_function.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Prints out the summary of the model}
\usage{
\method{summary}{linreg}(object)
}
\arguments{
\item{object}{A object of class "linreg"}
}
\description{
Prints the summary
}
