% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_function.R
\name{linreg}
\alias{linreg}
\title{Creating linear model}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{The equation of the linear model}

\item{data}{The dataset used when creating linear models}
}
\value{
The function generates several components of the linear model, including regression coefficients, fitted values and the residual variance
}
\description{
Linreg is a function which is fitting linear models
}
